[genetic tuning] Do not print confusing stats
Each generation includes two evalution phases (marked with *):
1) Benchmarking of candidates *
2) Breeding
3) Benchmarking of new candidates and selection *

Except for the first generation usually most of the candidates of
step 1) will have been evaluated during step 3) of the previous
generation.

This commit removed the "Jobs(Compiled,GPU) (0,0)/0" that was
printed during step 1) of almost every generation.

In theory it is possible not to have enough valid candidates to
form a new generation. In those cases step 1) will be run and
the progress message will be printed