Ensure input immutability
We have been talking about input immutability for a while in TC but it was enver actually implemented.
This adds a simple check to semantic analysis as well as a unit test to expect that input parameters are immutable.
Note that this is not foolproof because one may always abuse tmp/output tensors to be a view into a subportion of an input tensor but this is not intended to be commonplace.

In the future we may want to relax input immutability.