reduction_matcher.cc: isAlmostIdentityReduction: directly check isl::pw_aff
There is no need to check that the piecewise expression only has
a single cell.  Having multiple cells does not affect the nature
of being a reduction.
The purpose of this change is not to catch more cases since
the isl scheduler currently does not produce any piecewise
band members, but rather to simplify the code.