drop unused "mu wrappers"
Since these wrappers are unused, they should be removed as dead code.
They are apparently also dangerous (judging from the WARNING),
so the removal will make sure they don't get used in the future.
In particular, if these abstractions give the impression
that they could be used to modify parts of isl expressions
in place, then they should be killed with fire.
isl objects should be modified through high-level isl functions
and not by trying to tinker with the representation.

The MUPA wrapper (along with its dependencies) is still in use at this point
and is left in by this commit.
It will be removed later.
In particular, one use reveals a missing high-level isl function,
which should be added and used instead, one use disappears after
some simplification of the code and the remaining use can be
replaced by a call to get_union_pw_aff_list(), which will
also be added to isl.